+++ Classes/ViewHelpers/PageViewHelper.php
@@ -45,12 +45,14 @@ class PageViewHelper extends AbstractPDFViewHelper
     {
         parent::initializeArguments();
 
-        $this->registerArgument('autoPageBreak', 'boolean', '', false, $this->settings['page']['autoPageBreak']);
-        $this->registerArgument('margin', 'array', '', false, []);
-        $this->registerArgument('importPage', 'integer', '', false, $this->settings['page']['importPage']);
-        $this->registerArgument('importPageOnAutomaticPageBreak', 'boolean', '', false, $this->settings['page']['importPageOnAutomaticPageBreak']);
-        $this->registerArgument('orientation', 'string', '', false, $this->settings['page']['orientation']);
-        $this->registerArgument('format', 'string', '', false, $this->settings['page']['format']);
+        $this->registerArgument('autoPageBreak', 'boolean', 'If true a new page is added if the content exceeds the current page.', false, $this->settings['page']['autoPageBreak']);
+        $this->registerArgument('margin', 'array', 'The margin of this page.', false, []);
+        $this->registerArgument('importPage', 'integer', 'The page number to be imported as template.', false, $this->settings['page']['importPage']);
+        $this->registerArgument('importPageOnAutomaticPageBreak', 'boolean', 'If true the template is also applied if an automatic page break occurs.', false, $this->settings['page']['importPageOnAutomaticPageBreak']);
+        $this->registerArgument('orientation', 'string', 'The orientation of the page: portrait, landscape', false, $this->settings['page']['orientation']);
+        $this->registerArgument('format', 'string', 'The format of the page, e.g. A4', false, $this->settings['page']['format']);
+        $this->registerArgument('keepMargins', 'boolean', 'If true overwrites the default page margins with the current margins.', false, $this->settings['page']['keepMargins']);
+        $this->registerArgument('tableOfContentPage', 'boolean', 'If true the page will be rendered as a table of content page, e.g. it can be moved to the front.', false, $this->settings['page']['tableOfContentPage']);
     }
 
     /**
@@ -91,7 +93,7 @@ class PageViewHelper extends AbstractPDFViewHelper
         $this->getPDF()->SetMargins($this->arguments['margin']['left'], $this->arguments['margin']['top'], $this->arguments['margin']['right']);
         $this->getPDF()->SetAutoPageBreak($this->arguments['autoPageBreak'], $this->arguments['margin']['bottom']);
 
-        $this->getPDF()->AddPage($this->arguments['orientation'], $this->arguments['format']);
+        $this->getPDF()->AddPage($this->arguments['orientation'], $this->arguments['format'], $this->arguments['keepMargins'], $this->arguments['tableOfContentPage']);
 
         if ($hasImportedPage) {
             $this->getPDF()->useTemplate($template);
@@ -109,6 +111,10 @@ class PageViewHelper extends AbstractPDFViewHelper
             //no auto page break occurred or no element was rendered, we still need to set the header
             $this->setPageNeedsHeader(false);
             $this->getPDF()->renderHeader();
+        }
+
+        if ($this->arguments['tableOfContentPage']) {
+            $this->getPDF()->endTOCPage();
         }
 
         //reset default header and footer scope to document