+++ Tests/Functional/ExamplesTest.php
@@ -28,6 +28,8 @@ namespace Bithost\Pdfviewhelpers\Tests\Functional;
  *  This copyright notice MUST APPEAR in all copies of the script!
  * * */
 
+use GuzzleHttp\Client;
+
 /**
  * ExamplesTest
  *
@@ -64,6 +66,8 @@ class ExamplesTest extends AbstractFunctionalTest
 
         $this->assertContains('Avoid page break inside', $pages[4]->getText());
 
+        $this->validatePDF($output);
+
         //do not assert file equality, because the files generated on the server are not equal to the local files
         //comparing hashes of two locally generated files works however
         //$expectedHash = sha1(file_get_contents($this->getFixturePath('Examples/FullFeatureShowCase.pdf')));
@@ -90,6 +94,8 @@ class ExamplesTest extends AbstractFunctionalTest
         $this->assertContains('Esteban Marín, Markus Mächler', $text);
         $this->assertContains('Bithost GmbH', $text);
 
+        $this->validatePDF($output);
+
         //$expectedHash = sha1(file_get_contents($this->getFixturePath('Examples/BasicUsage.pdf')));
         //$actualHash = sha1($output);
         //$this->assertEquals($expectedHash, $actualHash);
@@ -114,8 +120,108 @@ class ExamplesTest extends AbstractFunctionalTest
         $this->assertContains('Here is the HTML header', $text);
         $this->assertContains('Lorem ipsum dolor sit amet', $text);
 
+        $this->validatePDF($output);
+
         //$expectedHash = sha1(file_get_contents($this->getFixturePath('Examples/ExtendExistingPDFs.pdf')));
         //$actualHash = sha1($output);
         //$this->assertEquals($expectedHash, $actualHash);
+    }
+
+    /**
+     * @test
+     */
+    public function testTableOfContent()
+    {
+        $this->setUpPage([$this->getFixturePath('Examples/TableOfContent.txt')]);
+
+        $output = $this->renderFluidTemplate($this->getFixturePath('Examples/TableOfContent.html'));
+        $pdf = $this->parseContent($output);
+        $pages = $pdf->getPages();
+
+        $this->assertCount(5, $pages);
+
+        $this->assertContains('HTML Table of content', $pages[0]->getText());
+        $this->assertContains('html-table-of-content-header', $pages[0]->getText());
+        $this->assertContains('html-table-of-content-footer', $pages[0]->getText());
+        $this->assertContains('LEVEL 0: Headline page 1', $pages[0]->getText());
+        $this->assertContains('LEVEL 0: Headline page 2', $pages[0]->getText());
+        $this->assertContains('LEVEL 1: Headline page 2, level 1', $pages[0]->getText());
+        $this->assertContains('LEVEL 0: Adding custom styled bookmark for a text', $pages[0]->getText());
+        $this->assertContains('LEVEL 0: Adding custom styled ADVANCED bookmark for a text', $pages[0]->getText());
+        $this->assertNotContains('Headline not in table of content', $pages[0]->getText());
+
+        $this->assertContains('Regular table of content', $pages[1]->getText());
+        $this->assertContains('regular-table-of-content-header', $pages[1]->getText());
+        $this->assertContains('regular-table-of-content-footer', $pages[1]->getText());
+        $this->assertContains('Headline page 1', $pages[1]->getText());
+        $this->assertContains('Headline page 2', $pages[1]->getText());
+        $this->assertContains('Headline page 2, level 1', $pages[1]->getText());
+        $this->assertContains('Adding custom styled bookmark for a text', $pages[1]->getText());
+        $this->assertContains('Adding custom styled ADVANCED bookmark for a text', $pages[1]->getText());
+        $this->assertNotContains('LEVEL 0', $pages[1]->getText());
+        $this->assertNotContains('LEVEL 1', $pages[1]->getText());
+        $this->assertNotContains('Headline not in table of content', $pages[1]->getText());
+
+        $this->assertContains('Headline page 1', $pages[2]->getText());
+        $this->assertContains('General Header', $pages[2]->getText());
+        $this->assertContains('General Footer', $pages[2]->getText());
+
+        $this->assertContains('Headline page 2', $pages[3]->getText());
+        $this->assertContains('Headline page 2, level 1', $pages[3]->getText());
+        $this->assertContains('Headline page 2, level 2', $pages[3]->getText());
+
+        $this->assertContains('Headline not in table of content', $pages[4]->getText());
+        $this->assertContains('Here is some text', $pages[4]->getText());
+        $this->assertContains('Headline not in table of content', $pages[4]->getText());
+
+        //Do not validate as TCPDF does not produce valid documents having bookmarks
+        //$this->validatePDF($output);
+    }
+
+    /**
+     * @test
+     */
+    public function testPDFA()
+    {
+        $this->setUpPage([$this->getFixturePath('Examples/PDFA.txt')]);
+
+        $output = $this->renderFluidTemplate($this->getFixturePath('Examples/PDFA.html'));
+
+        $this->validatePDF($output);
+    }
+
+    /**
+     * @param string $pdf
+     */
+    protected function validatePDF($pdf)
+    {
+        $client = new Client();
+        $response = $client->post('https://www.pdf-online.com/osa/validate.aspx', [
+            'headers' => [
+                'Accept' => 'application/json'
+            ],
+            'multipart' => [
+                [
+                    'name'     => 'file',
+                    'contents' => $pdf,
+                    'headers' => [
+                        'Content-Disposition' => 'form-data; name="file"; filename="document.pdf"',
+                        'Content-Type' => 'application/pdf',
+                    ]
+                ]
+            ]
+        ]);
+
+        $this->assertEquals(200, $response->getStatusCode());
+
+        $responseData = json_decode($response->getBody(), true);
+
+        $this->assertArrayHasKey('Result', $responseData);
+        $this->assertArrayHasKey('Details', $responseData);
+        $this->assertEquals(
+            'Document validated successfully.',
+            $responseData['Result'],
+            implode(PHP_EOL, $responseData['Details'])
+        );
     }
 }