+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog for TYPO3 CMS Extension pdfviewhelpers
 
+## 2.1.0 - April 30, 2019
+- Adds paragraph line feed to textual ViewHelpers, [#107](https://github.com/bithost-gmbh/pdfviewhelpers/issues/107) (Thanks [@emmemme](https://github.com/emmemme))
+- Adds image processing to ImageViewHelper, [#104](https://github.com/bithost-gmbh/pdfviewhelpers/issues/104)
+- Adds table of content ViewHelpers, [#103](https://github.com/bithost-gmbh/pdfviewhelpers/issues/103)
+- Updates TCPDF to version `6.2.26`
+- Updates FPDI to version `2.2.0`
+- Updates documentation
+
 ## 2.0.2 - December 1, 2018
 - Fixes import page on automatic page break not working, [#99](https://github.com/bithost-gmbh/pdfviewhelpers/issues/99) (Thanks [@liayn](https://github.com/liayn))
 - Updates documentation